Update meter_page : Seperated meter_animation widget
